pyproject.toml
[project]
name = "bot-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord.py>=2.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.30.0",
    "google-genai>=1.56.0",
]
[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
